refactor: use alias rather than build file for type export
