--- src/types/client.ts
@@ -1,3 +1,0 @@
-import { createClient } from '#sanity-client'
-
-export type SanityClient = ReturnType<typeof createClient>
