+++ test/unit/client.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, it, vi, expect } from 'vitest'
 import { $fetch } from 'ofetch'
-import { getQuery, createClient } from '../../src/runtime/client'
+import { getQuery, createClient } from '../../src/runtime/minimal-client'
 import { request as largeRequest } from './fixture/large-request.json'
 
 vi.mock('ofetch', () => ({
